separate operations test
